Added exploratory plots
I tried to see whether the data could be interpreted as a Poisson
process. The inter-arrival histograms as well as the counter for the
number of new IPs discovered do not provide evidence for a Poisson
process.

Using the log of the inter-arrival time for new IPs shows two
interesting mixtures in our data, but it cannot be used to identify an
IoT device.